Version2

+++ hook/delete.php
@@ -1,7 +1,12 @@
 <?php
 require_once '../tools/warframe.php';
 $session->is_auth();
-$form = new $_GET['model'];
-unset($_GET['model']);
-$form->delete($_GET['id']);
+
+if ( isset($_GET['model']) ) {
+    $form = new $_GET['model'];
+    unset($_GET['model']);
+    $form->delete($_GET['id']);
+}else {
+    Mixin\error('403');
+}
 ?>
